02-snowflake/requirements.txt
outerbounds
metaflow-snowflake
metaflow-dbt
dbt-snowflake